add data classes for rest: entity -> model -> dto

+++ sir-richard-boss-rest/src/main/java/com/sir/richard/boss/model/types/CarrierStatuses.java
@@ -64,7 +64,7 @@ public enum CarrierStatuses {
 		return annotation;
 	}
 
-	public static CarrierStatuses getValueById(int value) {
+	public static CarrierStatuses getValueById(Long value) {
 	
 		if (value == 0) {
 			return CarrierStatuses.UNKNOWN;
